fix: common text extension
